New script: list to Reminders
Inspired by iOS Drafts app. This is a script that will take a list that is on your clipboard and put every line as new reminder in Reminders. If the first line starts with `!` followed by the name of one of your lists, the default list (`Shoppinglist`) will not be used.