New Action: Booklet
Runs a bookmarklet in Google Chrome.

NOTE: You have to specify the name of your bookmarklets, and they have
to be at the root of the Bookmarks menu OR in a submenu named "More".
I'm too lazy to iterate the menu items right now. Feel welcomed to send
me a pull request.